Creating a new test to verify #1820
Test isn't implementing logic of #1820 yet, just replicating an existing test to ensure the test runner passes initially.

Created in a new test class to reduce changes of merge conflicts, but extracted out shared logic from existing test class.
